[UPDATE]local event

+++ app/src/main/java/com/dms/pmsandroid/feature/calendar/viewmodel/CalendarViewModel.kt
@@ -51,8 +51,6 @@ class CalendarViewModel(
     }
 
     fun loadSchedules() {
-
-
         val accessToken = sharedPreferenceStorage.getInfo("access_token")
         calendarApiImpl.scheduleApi(accessToken).subscribe { response ->
             if (response.isSuccessful) {
@@ -60,7 +58,6 @@ class CalendarViewModel(
                     eventDatabase.eventDao().deleteEvents()
                 }
                 parseEvents(response.body()!!)
-
             }
         }
     }
@@ -69,9 +66,9 @@ class CalendarViewModel(
         eventDatabase.eventDao().getLocalEvent().observeOn(AndroidSchedulers.mainThread())
             .subscribeOn(Schedulers.io()).subscribe { events ->
                 for (event in events) {
-                    val month = event.date.substring(4, 6).toInt()+1
+                    val month = (event.date.substring(4, 6).toInt())+1
                     val eventKey = EventKeyModel(month, event.date)
-                    val dotTypes = event.dot.split(",").map { it.trim() }
+                    val dotTypes = event.dot.split(",").map { it.trim().toInt() }
                     val eventModel = EventModel(event.event, dotTypes as ArrayList<Int>)
                     _events.value!![eventKey] = eventModel
                 }
